implemted Atlas Builder

+++ ConsoleApplication1/ConsoleApplication1/Sprite.cs
@@ -14,8 +14,8 @@ namespace ConsoleApplication1
         public string safeName;
         public BitmapImage image;
 
-        public float posX;
-        public float posY;
+        public int posX;
+        public int posY;
 
         public int width;
         public int height;
